Security: do not pass unescaped stuff to shell.
Previously, all the arguments were spewed back to the shell via perl's
backtick operator — and without any escaping, or even quoting it. That
is a security issue waiting to happen.

This changes it to avoid the shell entirely. It's easiest to use a
module (instead of dealing with the moving file descriptors around then
the pipe form of open); I picked IPC::Run3 which is relatively
lightweight. Since that rules out building a shell pipeline, could no
longer use "openssl x509" to extract just the cert; instead use a
regexp.

Note extracting the cert doesn't appear to be required, but not removing
it as possibly other versions of the module need it. I presume it was
there for a reason.